backend/requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
sqlalchemy==2.0.36
httpx==0.28.1
python-dotenv==1.1.1
pydantic==2.12.3
